+++ ebrium/__init__.py
@@ -1,3 +1,3 @@
 """ebrium - Font metrics normalization package."""
 
-__version__ = "3.0.1"
+__version__ = "3.1.0"